bug fixes:  - origin force no longer has an r^2 strength when r<1 because it CAUSED    instability in nested clusters  - collapsing and expanding of nested clusters should now work more reliably    due to more sensible recursive logic in those methods

+++ src/org/wilmascope/graph/Edge.java
@@ -230,9 +230,17 @@ public class Edge extends GraphElement {
     owner.remove(this);
     start.removeEdge(this);
     end.removeEdge(this);
-    view.delete();
+    if(view != null) {
+      view.delete();
+    } else {
+      System.out.println("hmmm... null view while deleting edge, was it collapsed?");
+    }
     view = null;
-    layout.delete();
+    if(layout != null) {
+      layout.delete();
+    } else {
+      System.out.println("hmmm... null layout while deleting edge, was it collapsed?");
+    }
     layout = null;
     recalculateMultiEdgeOffsets();
   }
